photos endpoint and docs done

+++ models/comment.go
@@ -8,10 +8,10 @@ import (
 )
 
 type Comment struct {
-	ID        uint64    `gorm:"primarykey" json:"id"`
-	UserID    uint64    `json:"user_id" form:"user_id"`
+	ID        uint    `gorm:"primarykey" json:"id"`
+	UserID    uint    `json:"user_id" form:"user_id"`
 	User      User      `json:"user" form:"user"`
-	PhotoID   uint64    `json:"photo_id" form:"photo_id"`
+	PhotoID   uint    `json:"photo_id" form:"photo_id"`
 	Photo     Photo     `json:"photo" form:"photo"`
 	Message   string    `gorm:"not null" json:"message" form:"message" valid:"required~message is required"`
 	CreatedAt time.Time `json:"created_at"`
